requirements.txt
customtkinter
csv
os
bcrypt
pillow